+++ CloudShell.Hosting/Hosting/CloudShellHostApplicationBuilderExtensions.cs
@@ -9,6 +9,7 @@ using CloudShell.Hosting.Observability;
 using CloudShell.Hosting.ResourceManager;
 using CloudShell.Hosting.Shell;
 using CoreShell;
+using CoreShell.Blazor;
 using CoreShell.Composition;
 using CoreShell.Composition.Blazor;
 using Microsoft.AspNetCore.Components.Authorization;
@@ -83,9 +84,8 @@ public static class CloudShellHostApplicationBuilderExtensions
         builder.Services.AddRazorComponents()
             .AddInteractiveServerComponents();
         builder.Services.AddCloudShellUiComposition();
+        builder.Services.AddCoreShellBlazor();
         builder.Services.TryAddSingleton<ShellHostContext>();
-        builder.Services.TryAddSingleton<ICoreShellContentResolver, BlazorCoreShellContentResolver>();
-        builder.Services.TryAddSingleton<ICoreShellLayoutResolver, BlazorCoreShellLayoutResolver>();
         builder.Services.TryAddSingleton<CoreShellModuleCatalog>();
         builder.Services.TryAddSingleton<ICoreShellPageService>(
             serviceProvider => serviceProvider.GetRequiredService<CoreShellModuleCatalog>());